Moved function from status to finbusinessdata and added new pool function
